Feat : add clipboard
